Some Sensor and Keyboard changes mainly dealing with Windows Phone development.
The Keyboard class now runs in a Windows Phone environment allowing you
to take input from a phone with a hardware keyboard. You now also have
access to the Motion class by enabling it through the Windows Phone
class. This gives access to all sensors on the phone and automatically
updates them and keeps data held if its enabled.