Finally decided it was time to put my latest work into the project. Check extended for more details.
Added two more projects and an SLN to make it easy to open all projects
at once. Graphics has some slight work done in it but it is still being
worked on daily. The optional stuff is mainly where I intend to keep any
project not directly related to a library.