Updates all around to the GUI system.
Added a lot of new things into the GUI system. Renamed GuiRenderer to
GuiHandler and it now handles Updating and Drawing our Gui elements.
Added somre more Interfaces and updated Label a bit. Next up I will be
returning to the Input engine for a bit to get it ready for easy access
Gui things.